Added --dump option to server to write traffic to a file
The --dump option makes it easy to inspect network traffic
even when server and client are running on the same host
and Wireshark can't be used to capture the packets.
It writes out the packets in a format similar to Wireshark's
hexdump. It can then be processed by the parse.py script.