Add first version of authcode handler
Authcodes will be used to allow someone to claim the same
username that he has used on Hirez' servers. He will be sent
the authcode when his identity has been verified and can
then use this on the server to claim his username and set
his password.

The getauthcode.py script can be used to request an authcode
for a specified username. The server doesn't store this code
yet at the moment, so it's useless at the moment.