Added script for capturing some game server packets
This script connects to the game server and starts replaying
some recorded packets. After these packets it keeps the server
talking.

All server packets are written out to the file serverpacketbindump.txt.
Each line in this file contains the size of a packet and the content
as a little-endian bitstream.

A separate script (similar to parse.py) will be written to read the
dump and format it for as far as we know the structure.