Replaced hexdump2bindump with carrays2bindump
Hexdumps don't contain information about where one packet
ends and the next one begins in case the number of bytes
in a packet is a multiple of 16. The C Arrays export of
Wireshark doesn't have this problem, so we'll use that.
Bindumps created with the hexdump script grouped multiple
packets into one bigger one, so regenerate your bindumps.