First version of module that parses UDK data into objects
The purpose of this module is that it can both be used to
parse incoming packets and construct outgoing packets.
Once that works it can be integrated into the game client
to get rid of recorded responses and to have it report
on relevant game events such as team switches and match
time.