Fix problem when running client/login/game server on one host
Now both the private and the global IP address of each player
and game server is tracked by the login server and the server
IP advertised to the player depends on whether the game server
is local to the player.

The only case that we'll never get to work is when the player
and game server are on the same LAN, but not the same host and
the game server firewall blocks the player. Then the user will
just have to manually add a firewall rule to allow connections
from the player. It's impossible to automate this because we
don't have software running on the client and therefore cannot
determine his local IP address.