Implement immediate blocking of individual clients on the same IP
Now when the game server opens up the firewall to allow a player in,
it passes the player ID. For each player ID a port will be stored as
soon as the first packet that matches the IP is seen. This allows us
to later on stop forwarding packets for this IP/port combination
when the player is removed from the server.

The assumption is that each client will use only one source port when
communicating with the game server. Relying on this allows us to
block individual players even if there are other players on the server
with the same IP.