Always disconnect a player when an exception occurs
Now when an exception occurs players are no longer kept
in the player list and will no longer have trouble logging
in again.

This handles the case where the exception occurs inside
the connectionreader. This is the most likely place where
it will happen, because it processes data created by the
client (as opposed to the connectionwriter that handles
data created within this program).