Fix problem with half-open TCP connections from clients
If a client lost network connectivity, the TCP connection
from the point of view of the login server would remain
open and the player would not be removed from the server.
This would make it impossible to log in with the same
account later, because the account was still in use.

Now the 01c8 messages that a client sends every 10 seconds
are used to see if a client is still present. If no such
messages have been received for a whole minute, the client
is disconnected and removed from the server.