Add warning about upgrading to master
Upgrading to master works, but downgrading from there to an older release doesn't.
This is because the version detected by the script when running on master is the
last release and it will consider any updates in master as locally modified files.
Consequently it will keep the versions from master when going to a different release
if they are the same in the detected release and the target release. The result
is likely a version that doesn't work.