Extend duration of map vote to ServerSettings.EndMatchWaitTime
The vote will last for either EndMatchWaitTime seconds or for
the time it takes to start up a new server, whichever is longer.
After the vote it will still take a few more seconds for the new
server to load the selected map before players are sent to the
new server and get the loading screen.